Se agrega archivo de usuarios en carpeta files y metodo LeerYEscribirUsuarios

+++ src/tpTurismoEnLaTierraMedia/CostoComparator.java
@@ -5,6 +5,6 @@ public class CostoComparator implements Comparator<Atraccion>{
 	
 	@Override
 	public int compare(Atraccion o1, Atraccion o2) {
-		return Integer.compare(o1.getCosto(), o2.getCosto());	
+		return Double.compare(o1.getCosto(), o2.getCosto());	
 	}
 }
